refactor(MovieAlarmSetter): 영화 알림 설정에 대한 기능을 MovieAlarmSetter에게 위임
